Improved friendly collision code
Minions now match the speed of the minion on front of them when they
collide. This stops the crappy jittering of the minion at the back if it
is faster than the one in front. Once the leading minion is no longer in
front of the trailing minion, the trailing minion will return to its
normal speed